+++ android_app/app/src/main/java/com/bh/fittingsimulator/MainActivity.java
@@ -25,7 +25,7 @@ public class MainActivity extends AppCompatActivity {
 
     private GLSurfaceView mGLView;
     private FrameLayout testLayout;
-    private double[] arr;
+    private float[] arr;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -72,13 +72,13 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-        arr=new double[]{0,0,0,0,0};
+        arr=new float[]{0,0,0,0,0};
         Intent intent = getIntent();
-        arr[0] = intent.getDoubleExtra("shoulder",0);
-        arr[1] = intent.getDoubleExtra("arm",0);
-        arr[2] = intent.getDoubleExtra("chest",0);
-        arr[3] = intent.getDoubleExtra("arm_width",0);
-        arr[4] = intent.getDoubleExtra("total_len",0);
+        arr[0] = intent.getFloatExtra("shoulder",0);
+        arr[1] = intent.getFloatExtra("arm",0);
+        arr[2] = intent.getFloatExtra("chest",0);
+        arr[3] = intent.getFloatExtra("arm_width",0);
+        arr[4] = intent.getFloatExtra("total_len",0);
 
 
     }
