+++ android_app/app/src/main/java/com/bh/fittingsimulator/SelectClothesActivity.java
@@ -31,7 +31,7 @@ public class SelectClothesActivity extends AppCompatActivity {
     private IntentIntegrator qrScan;
     private String[] clothes_data;
 
-    private double[] arr=new double[6];
+    private float[] arr=new float[6];
 
     private EditText top_shoulder, top_chest, top_arm, top_arm_width,top_total_len;
 
@@ -101,15 +101,15 @@ public class SelectClothesActivity extends AppCompatActivity {
 
                 //상의 치수 변수들 -> 서버로 보내기
                 try{
-                    arr[0]=Double.parseDouble(top_shoulder.getText().toString());//더블형
+                    arr[0]=Float.parseFloat(top_shoulder.getText().toString());//더블형
 
-                    arr[1]=Double.parseDouble(top_arm.getText().toString());//더블형
+                    arr[1]=Float.parseFloat(top_arm.getText().toString());//더블형
 
-                    arr[2]=Double.parseDouble(top_chest.getText().toString());//더블형
+                    arr[2]=Float.parseFloat(top_chest.getText().toString());//더블형
 
-                    arr[3]=Double.parseDouble(top_arm_width.getText().toString());//더블형
+                    arr[3]=Float.parseFloat(top_arm_width.getText().toString());//더블형
 
-                    arr[4]=Double.parseDouble(top_total_len.getText().toString());//더블형
+                    arr[4]=Float.parseFloat(top_total_len.getText().toString());//더블형
 
                 } catch (NumberFormatException e) {
                     e.printStackTrace();
